work in progress
Former-commit-id: 381397c70e7d4aca569de488ce08d6aad9ff8bd7
Former-commit-id: e483d212c6fe807b70900d61a74d7383e3c25bd9

+++ src/wid_button.c
@@ -23,8 +23,8 @@ wid_button_internal (const char *text, float x, float y, fontp font,
                      uint32_t text_scale)
 {
     widp wid_button_window;
-    uint32_t maxw = 0;
-    uint32_t maxh = 0;
+    double maxw = 0;
+    double maxh = 0;
     int32_t mx, my;
 
     /*
